requirement.txt
Django>=3.2
psycopg2-binary
Pillow
requests
argon2-cffi
djangorestframework